fixed all warnings except some in the HLSL files, also changed default initial background color from blue to red, and we are able to run in release

+++ Game/Source/Menu.cpp
@@ -152,7 +152,7 @@ void Menu::Render(Direct3D* _direct3D)
 	_direct3D->Render(menu_background, XMLoadFloat4x4(&camera->GetViewMatrix()));
 
 	// Render menu buttons
-	for (int i = 0; i < menuButtons.size(); i++)
+	for (size_t i = 0; i < menuButtons.size(); i++)
 	{
 		_direct3D->Render(menu_background, XMLoadFloat4x4(&camera->GetViewMatrix()));
